pyproject.toml
[tool.poetry]
name = "midi-web-app"
version = "0.1.0"
description = ""
authors = ["AydenClay96 <AydenClay1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pygame-ce = "^2.5.1"
black = "^24.8.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
